refactor(admin): divide into submodules
