+++ src/admin/dashboard/dashboard.const.ts
@@ -1,0 +1,5 @@
+import { RouteParts } from '../../constants';
+
+export const DASHBOARD_PATH = {
+	DASHBOARD: `/${RouteParts.Admin}`,
+};
