+++ src/admin/dashboard/views/index.ts
@@ -1,0 +1,3 @@
+import Dashboard from './Dashboard';
+
+export { Dashboard };
