+++ src/admin/menu/menu.gql.ts
