+++ src/admin/menu/views/MenuDetail.scss
