+++ src/admin/menu/views/MenuDetail.tsx
@@ -10,14 +10,10 @@ import { navigate } from '../../../shared/helpers';
 import { ApolloCacheManager } from '../../../shared/services/data-service';
 import toastService, { TOAST_TYPE } from '../../../shared/services/toast-service';
 
-import { ADMIN_PATH } from '../../admin.const';
-import {
-	DELETE_MENU_ITEM,
-	GET_MENU_ITEMS_BY_PLACEMENT,
-	UPDATE_MENU_ITEM_BY_ID,
-} from '../../admin.gql';
-import { MenuItem } from '../../admin.types';
-import { AdminLayout, AdminLayoutActions, AdminLayoutBody } from '../../layouts';
+import { AdminLayout, AdminLayoutActions, AdminLayoutBody } from '../../shared/layouts';
+import { MENU_PATH } from '../menu.const';
+import { DELETE_MENU_ITEM, GET_MENU_ITEMS_BY_PLACEMENT, UPDATE_MENU_ITEM_BY_ID } from '../menu.gql';
+import { MenuItem } from '../menu.types';
 
 import './MenuDetail.scss';
 
@@ -160,7 +156,7 @@ const MenuDetail: FunctionComponent<MenuDetailProps> = ({ history, match }) => {
 		return (
 			<AdminLayout
 				className="c-menu-detail"
-				navigateBack={() => handleNavigate(ADMIN_PATH.MENU)}
+				navigateBack={() => handleNavigate(MENU_PATH.MENU)}
 				pageTitle={startCase(menuId)}
 			>
 				<AdminLayoutBody>
@@ -183,7 +179,7 @@ const MenuDetail: FunctionComponent<MenuDetailProps> = ({ history, match }) => {
 											<Button
 												icon="edit-2"
 												onClick={() =>
-													handleNavigate(ADMIN_PATH.MENU_EDIT, {
+													handleNavigate(MENU_PATH.MENU_EDIT, {
 														menu: menuId,
 														id: String(item.id),
 													})
@@ -207,7 +203,7 @@ const MenuDetail: FunctionComponent<MenuDetailProps> = ({ history, match }) => {
 								icon="plus"
 								label="Voeg een item toe"
 								onClick={() =>
-									handleNavigate(ADMIN_PATH.MENU_CREATE, {
+									handleNavigate(MENU_PATH.MENU_CREATE, {
 										menu: menuId,
 									})
 								}
@@ -227,11 +223,7 @@ const MenuDetail: FunctionComponent<MenuDetailProps> = ({ history, match }) => {
 						label="Opslaan"
 						onClick={() => handleSave(refetch)}
 					/>
-					<Button
-						label="Annuleer"
-						onClick={() => handleNavigate(ADMIN_PATH.MENU)}
-						type="tertiary"
-					/>
+					<Button label="Annuleer" onClick={() => handleNavigate(MENU_PATH.MENU)} type="tertiary" />
 				</AdminLayoutActions>
 			</AdminLayout>
 		);
