+++ src/admin/menu/views/MenuOverview.tsx
@@ -7,10 +7,10 @@ import { Button, ButtonToolbar, Table } from '@viaa/avo2-components';
 import { DataQueryComponent } from '../../../shared/components';
 import { navigate } from '../../../shared/helpers';
 
-import { ADMIN_PATH, MENU_OVERVIEW_TABLE_COLS } from '../../admin.const';
-import { GET_MENUS } from '../../admin.gql';
-import { MenuItem, MenuOverviewTableCols } from '../../admin.types';
-import { AdminLayout } from '../../layouts';
+import { AdminLayout } from '../../shared/layouts';
+import { MENU_OVERVIEW_TABLE_COLS, MENU_PATH } from '../menu.const';
+import { GET_MENUS } from '../menu.gql';
+import { MenuItem, MenuOverviewTableCols } from '../menu.types';
 
 interface MenuOverviewProps extends RouteComponentProps {}
 
@@ -24,14 +24,14 @@ const MenuOverview: FunctionComponent<MenuOverviewProps> = ({ history }) => {
 					<ButtonToolbar>
 						<Button
 							icon="list"
-							onClick={() => navigate(history, ADMIN_PATH.MENU_DETAIL, { menu: rowData.placement })}
+							onClick={() => navigate(history, MENU_PATH.MENU_DETAIL, { menu: rowData.placement })}
 							size="small"
 							title="Bekijk alle navigatie items"
 							type="tertiary"
 						/>
 						<Button
 							icon="plus"
-							onClick={() => navigate(history, ADMIN_PATH.MENU_CREATE, { menu: rowData.placement })}
+							onClick={() => navigate(history, MENU_PATH.MENU_CREATE, { menu: rowData.placement })}
 							size="small"
 							title="Voeg een navigatie item toe"
 							type="tertiary"
