+++ src/admin/menu/views/index.ts
@@ -1,0 +1,5 @@
+import MenuDetail from './MenuDetail';
+import MenuEdit from './MenuEdit';
+import MenuOverview from './MenuOverview';
+
+export { MenuDetail, MenuEdit, MenuOverview };
