+++ src/admin/shared/components/IconPicker/IconPicker.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/settings/colors';
-@import '../../../styles/settings/variables';
+@import '../../../../styles/settings/colors';
+@import '../../../../styles/settings/variables';
 
 .c-icon-picker {
 	.react-select__control {
