+++ src/admin/shared/components/Sidebar/Sidebar.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/settings/colors';
+@import '../../../../styles/settings/colors';
 
 /* ==========================================================================
    Component: Sidebar
