--- src/admin/views/index.ts
@@ -1,6 +1,0 @@
-import Dashboard from './Dashboard';
-import MenuDetail from './Menu/MenuDetail';
-import MenuEdit from './Menu/MenuEdit';
-import MenuOverview from './Menu/MenuOverview';
-
-export { Dashboard, MenuDetail, MenuEdit, MenuOverview };
